refactor: Separate PlaylistController buttons into individual files
