+++ src/components/features/playlist-manager/copy-button/index.ts
@@ -1,0 +1,1 @@
+export * from "./CopyButton";
