+++ src/components/features/playlist-manager/delete-button/index.ts
@@ -1,0 +1,1 @@
+export * from "./DeleteButton";
